File renaming and refactoring
