+++ src/routes/task.routes.ts
@@ -6,7 +6,7 @@ import {
   getTasksForUser,
   createTask,
   deleteTask,
-} from "../controllers/TaskController";
+} from "../controllers/task.controller";
 
 const router = express.Router();
