+++ src/routes/team.routes.ts
@@ -1,0 +1,14 @@
+import express from "express";
+import {
+  createTeam,
+  getAllTeams,
+  getTeamById,
+} from "../controllers/team.controller";
+
+const router = express.Router();
+
+router.post("/teams", createTeam);
+router.get("/teams", getAllTeams);
+router.get("/teams/:id", getTeamById);
+
+export default router;
