--- src/routes/teamRoutes.ts
@@ -1,10 +1,0 @@
-import express from 'express';
-import { createTeam, getAllTeams, getTeamById } from '../controllers/TeamController';
-
-const router = express.Router();
-
-router.post('/teams', createTeam);
-router.get('/teams', getAllTeams);
-router.get('/teams/:id', getTeamById);
-
-export default router;
